Rename package. Remove unnecessary interfaces
