+++ core/src/main/java/io/micronaut/core/util/locale/AbstractLocaleResolver.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.core.util.localeresolution;
+package io.micronaut.core.util.locale;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 import io.micronaut.core.util.LocaleResolver;
@@ -33,10 +33,10 @@ public abstract class AbstractLocaleResolver<T> implements LocaleResolver<T> {
 
     /**
      *
-     * @param localeResolutionConfiguration Locale Resolution configuration
+     * @param defaultLocale The locale to use if one cannot be found.
      */
-    public AbstractLocaleResolver(LocaleResolutionConfiguration localeResolutionConfiguration) {
-        this.defaultLocale = localeResolutionConfiguration.getDefaultLocale();
+    public AbstractLocaleResolver(Locale defaultLocale) {
+        this.defaultLocale = defaultLocale;
     }
 
     @Override
