--- core/src/main/java/io/micronaut/core/util/localeresolution/DefaultLocaleResolutionConfiguration.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2017-2020 original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.micronaut.core.util.localeresolution;
-
-import edu.umd.cs.findbugs.annotations.NonNull;
-
-import java.util.Locale;
-
-/**
- * Default Locale resolution configuration.
- *
- * @author Sergio del Amo
- * @since 2.3.0
- */
-public interface DefaultLocaleResolutionConfiguration {
-
-    /**
-     * @return The locale to be used if one cannot be resolved.
-     */
-    @NonNull
-    Locale getDefaultLocale();
-}
