+++ http-server/src/main/java/io/micronaut/http/server/binding/LocaleArgumentBinder.java
@@ -19,7 +19,7 @@ import io.micronaut.core.convert.ArgumentConversionContext;
 import io.micronaut.core.type.Argument;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.http.bind.binders.TypedRequestArgumentBinder;
-import io.micronaut.http.server.util.localeresolution.HttpLocaleResolver;
+import io.micronaut.http.server.util.locale.HttpLocaleResolver;
 
 import javax.inject.Singleton;
 import java.util.Locale;
