+++ http-server/src/main/java/io/micronaut/http/server/util/locale/HttpLocaleResolutionConfiguration.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.http.server.util.localeresolution;
+package io.micronaut.http.server.util.locale;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
-import io.micronaut.core.util.localeresolution.LocaleResolutionConfiguration;
+import io.micronaut.core.util.locale.LocaleResolutionConfiguration;
 
 import java.util.Optional;
