+++ http-server/src/test/groovy/io/micronaut/http/server/util/locale/FixedLocaleResolverSpec.groovy
@@ -1,9 +1,10 @@
-package io.micronaut.http.server.util.localeresolution
+package io.micronaut.http.server.util.locale
 
 import io.micronaut.context.ApplicationContext
 import spock.lang.Specification
 
-class HttpFixedLocaleResolverSpec extends Specification {
+class FixedLocaleResolverSpec extends Specification {
+
     void "bean of type HttpFixedLocaleResolver is not enabled by default"() {
         given:
         ApplicationContext context = ApplicationContext.run()
