+++ http-server/src/test/groovy/io/micronaut/http/server/util/locale/HttpFixedLocaleResolverSpec.groovy
@@ -1,10 +1,9 @@
-package io.micronaut.http.server.util.localeresolution
+package io.micronaut.http.server.util.locale
 
 import io.micronaut.context.ApplicationContext
 import spock.lang.Specification
 
-class FixedLocaleResolverSpec extends Specification {
-
+class HttpFixedLocaleResolverSpec extends Specification {
     void "bean of type HttpFixedLocaleResolver is not enabled by default"() {
         given:
         ApplicationContext context = ApplicationContext.run()
